model improvements
