+++ rl4co/models/zoo/hetgnn/encoder.py
@@ -8,6 +8,7 @@ from einops import einsum
 from torch import Tensor
 
 from rl4co.models.nn.env_embeddings import env_init_embedding
+from rl4co.models.nn.ops import Normalization
 
 
 class HetGNNLayer(nn.Module):
@@ -24,7 +25,7 @@ class HetGNNLayer(nn.Module):
         self.scale = 1 / math.sqrt(embed_dim)
 
     def forward(
-        self, self_emb: Tensor, other_emb: Tensor, edge_emb: Tensor, mask: Tensor
+        self, self_emb: Tensor, other_emb: Tensor, edge_emb: Tensor, edges: Tensor
     ):
         bs, n_rows, _ = self_emb.shape
 
@@ -35,9 +36,9 @@ class HetGNNLayer(nn.Module):
         ee = einsum(edge_emb, self.edge_attn, "b m o e, e one -> b m o") * self.scale
 
         # element wise multiplication similar to broadcast column logits over rows with masking
-        ec_expanded = einsum(mask, ec, "b m o, b o -> b m o")
+        ec_expanded = einsum(edges, ec, "b m o, b o -> b m o")
         # element wise multiplication similar to broadcast row logits over cols with masking
-        er_expanded = einsum(mask, er, "b m o, b m -> b m o")
+        er_expanded = einsum(edges, er, "b m o, b m -> b m o")
 
         # adding the projections of different node types and edges together (equivalent to first concat and then project)
         # (bs, n_rows, n_cols)
@@ -49,12 +50,12 @@ class HetGNNLayer(nn.Module):
         # (bs, n_ma, n_ops + 1)
         mask = torch.cat(
             (
-                mask == 1,
+                edges == 1,
                 torch.full(
                     size=(bs, n_rows, 1),
                     dtype=torch.bool,
                     fill_value=True,
-                    device=mask.device,
+                    device=edges.device,
                 ),
             ),
             dim=-1,
@@ -78,6 +79,24 @@ class HetGNNLayer(nn.Module):
         return hidden
 
 
+class HetGNNBlock(nn.Module):
+    def __init__(self, embed_dim) -> None:
+        super().__init__()
+        self.norm1 = Normalization(embed_dim, normalization="batch")
+        self.norm2 = Normalization(embed_dim, normalization="batch")
+        self.hgnn1 = HetGNNLayer(embed_dim)
+        self.hgnn2 = HetGNNLayer(embed_dim)
+
+    def forward(self, x1, x2, edge_emb, edges):
+        h1 = self.hgnn1(x1, x2, edge_emb, edges)
+        h1 = self.norm1(h1 + x1)
+
+        h2 = self.hgnn2(x2, x1, edge_emb.transpose(1, 2), edges.transpose(1, 2))
+        h2 = self.norm2(h2 + x2)
+
+        return h1, h2
+
+
 class HetGNNEncoder(nn.Module):
     def __init__(
         self,
@@ -98,8 +117,7 @@ class HetGNNEncoder(nn.Module):
         self.edge_weights_key = edge_weights_key
 
         self.num_layers = num_layers
-        self.row_emb = nn.ModuleList([HetGNNLayer(embed_dim) for _ in range(num_layers)])
-        self.col_emb = nn.ModuleList([HetGNNLayer(embed_dim) for _ in range(num_layers)])
+        self.layers = nn.ModuleList([HetGNNBlock(embed_dim) for _ in range(num_layers)])
 
     def forward(self, td):
         edges = td[self.edge_key]
@@ -108,13 +126,7 @@ class HetGNNEncoder(nn.Module):
         assert row_emb.size(1) == n_rows, "incorrect number of row embeddings"
         assert col_emb.size(1) == n_cols, "incorrect number of column embeddings"
 
-        for layer in range(self.num_layers):
-            row_emb_ = self.row_emb[layer](row_emb, col_emb, edge_emb, edges)
-            col_emb_ = self.row_emb[layer](
-                col_emb, row_emb, edge_emb.transpose(1, 2), edges.transpose(1, 2)
-            )
-
-            row_emb = row_emb_ + row_emb
-            col_emb = col_emb_ + col_emb
+        for layer in self.layers:
+            row_emb, col_emb = layer(row_emb, col_emb, edge_emb, edges)
 
         return (row_emb, col_emb), None
